# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "repo": {
            "title": "Repo",
            "type": "string"
          }
        },
        "required": [
          "repo"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "DiskUsageOut": {
        "description": "Wire mirror of `models.host_stats.DiskUsage` (the OpenAPI artifact is\nthe contract, so the shape is declared here rather than inferred).",
        "properties": {
          "avail_bytes": {
            "default": 0,
            "title": "Avail Bytes",
            "type": "integer"
          },
          "mount": {
            "title": "Mount",
            "type": "string"
          },
          "size_bytes": {
            "default": 0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "used_bytes": {
            "default": 0,
            "title": "Used Bytes",
            "type": "integer"
          }
        },
        "required": [
          "mount"
        ],
        "title": "DiskUsageOut",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The `{\"error\": {...}}` wire envelope every failure response that\nactually returns it uses (`terminals.py`, `setup.py`, the `app.py`\nmiddleware). Declared only on routes that return this exact shape \u2014\n`pairing.py`'s 403 returns `{\"detail\": ...}` instead (data-model.md \u00a72).",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorOut"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "remediation": {
            "title": "Remediation",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "remediation"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HostStatsResponse": {
        "description": "Wire mirror of `models.host_stats.HostStats`: a live snapshot, no\ntime series. ``temps`` is empty on hosts without sensors (VMs/containers);\nthe console hides the card entirely then.",
        "properties": {
          "cpu_count": {
            "default": 0,
            "title": "Cpu Count",
            "type": "integer"
          },
          "cpu_used_pct": {
            "default": 0.0,
            "title": "Cpu Used Pct",
            "type": "number"
          },
          "disks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DiskUsageOut"
            },
            "title": "Disks",
            "type": "array"
          },
          "load_1": {
            "default": 0.0,
            "title": "Load 1",
            "type": "number"
          },
          "load_15": {
            "default": 0.0,
            "title": "Load 15",
            "type": "number"
          },
          "load_5": {
            "default": 0.0,
            "title": "Load 5",
            "type": "number"
          },
          "mem_available": {
            "default": 0,
            "title": "Mem Available",
            "type": "integer"
          },
          "mem_total": {
            "default": 0,
            "title": "Mem Total",
            "type": "integer"
          },
          "mem_used": {
            "default": 0,
            "title": "Mem Used",
            "type": "integer"
          },
          "swap_total": {
            "default": 0,
            "title": "Swap Total",
            "type": "integer"
          },
          "swap_used": {
            "default": 0,
            "title": "Swap Used",
            "type": "integer"
          },
          "temps": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TempReadingOut"
            },
            "title": "Temps",
            "type": "array"
          },
          "uptime_s": {
            "default": 0.0,
            "title": "Uptime S",
            "type": "number"
          }
        },
        "title": "HostStatsResponse",
        "type": "object"
      },
...
      "JobAcceptedResponse": {
        "description": "202: the host detached a job; poll `GET /hosts/{id}/jobs/{job_id}`.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "kind",
          "project"
        ],
        "title": "JobAcceptedResponse",
        "type": "object"
      },
...
      "JobState": {
        "description": "Lifecycle state of a detached ``remo-host`` job.",
        "enum": [
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobStatusResponse": {
        "properties": {
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "finished_at": {
            "default": "",
            "title": "Finished At",
            "type": "string"
          },
          "log_tail": {
            "default": "",
            "title": "Log Tail",
            "type": "string"
          },
          "started_at": {
            "default": "",
            "title": "Started At",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          }
        },
        "required": [
          "state"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
...
      "ProjectDeletedResponse": {
        "properties": {
          "deleted": {
            "default": true,
            "title": "Deleted",
            "type": "boolean"
          },
          "project": {
            "title": "Project",
            "type": "string"
          }
        },
        "required": [
          "project"
        ],
        "title": "ProjectDeletedResponse",
        "type": "object"
      },
...
      "RebuildProjectRequest": {
        "properties": {
          "no_cache": {
            "default": false,
            "title": "No Cache",
            "type": "boolean"
          }
        },
        "title": "RebuildProjectRequest",
        "type": "object"
      },
...
      "TempReadingOut": {
        "description": "Wire mirror of `models.host_stats.TempReading`.",
        "properties": {
          "celsius": {
            "title": "Celsius",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "label",
          "celsius"
        ],
        "title": "TempReadingOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Remo Web Session Interface",
    "version": "0.1.0",
    "x-generated-by": "GENERATED FILE -- do not hand-edit. Regenerate with: uv run python scripts/export_openapi.py. See docs/maintaining-generated-types.md."
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hosts/{instance_id}/jobs/{job_id}": {
      "get": {
        "description": "Poll a detached job's status + log tail (console polls at 2s).",
        "operationId": "get_job_status_route_api_v1_hosts__instance_id__jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Get Job Status Route"
      }
    },
    "/api/v1/hosts/{instance_id}/projects": {
      "post": {
        "description": "Clone a repository into the host's ``projects_root`` (detached job).",
        "operationId": "create_project_api_v1_hosts__instance_id__projects_post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Create Project"
      }
    },
    "/api/v1/hosts/{instance_id}/projects/{project}": {
      "delete": {
        "description": "Delete a project (synchronous on the host: session + containers + dir).",
        "operationId": "delete_project_route_api_v1_hosts__instance_id__projects__project__delete",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDeletedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Delete Project Route"
      }
    },
    "/api/v1/hosts/{instance_id}/projects/{project}/rebuild": {
      "post": {
        "description": "Rebuild a project's devcontainer (detached job).",
        "operationId": "rebuild_project_api_v1_hosts__instance_id__projects__project__rebuild_post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RebuildProjectRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Rebuild Project"
      }
    },
    "/api/v1/hosts/{instance_id}/stats": {
      "get": {
        "description": "`GET /api/v1/hosts/{instance_id}/stats` -- live host statistics snapshot.\n\nUngated by design: the same trust level as `GET /hosts` (it discloses\nload/mem/temps to anyone who can reach the service \u2014 a documented\naccepted risk for this home-lab tool). Polling is coalesced by\n:class:`HostStatsCache` to at most one SSH call per TTL per host.",
        "operationId": "get_host_stats_route_api_v1_hosts__instance_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Get Host Stats Route"
      }
    },
...
  }
}